--- hono.ts
@@ -1,9 +1,0 @@
-import { Hono } from 'hono'
-import type { Session, User } from 'lucia'
-
-export const app = new Hono<{
-  Variables: {
-    user: User | null
-    session: Session | null
-  }
-}>()
