+++ lib/signatures.ts
